farmer_bot_final/requirements.txt
nltk
pickle
numpy
keras.models